refactor: factor out add_gossip_peer_from_header()
Also don't even add the peer to SQL if realtime is disabled.
